requirements.txt
argparse~=1.4.0
opencv-python~=4.5.1.48
numpy~=1.21.4
torchvision~=0.8.2
pillow~=8.0.1
sklearn~=0.0
scikit-learn~=0.24.0
tqdm~=4.56.0
torchsummary~=1.5.1
pandas~=1.3.4
ray~=1.8.0